Live casting system, inspired by @djarek 's one.
