+++ src/const.h
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -548,7 +548,10 @@ enum ReloadTypes_t : uint8_t  {
 
 static constexpr int32_t CHANNEL_GUILD = 0x00;
 static constexpr int32_t CHANNEL_PARTY = 0x01;
+static constexpr int32_t CHANNEL_CAST = 0xFF;
 static constexpr int32_t CHANNEL_PRIVATE = 0xFFFF;
+
+static constexpr auto CHANNEL_CAST_STR = "Live Cast";
 
 //Reserved player storage key ranges;
 //[10000000 - 20000000];
