+++ src/game.cpp
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -1820,6 +1820,11 @@ void Game::playerCloseChannel(uint32_t playerId, uint16_t channelId)
 		return;
 	}
 
+	if (channelId == CHANNEL_CAST) {
+		player->stopLiveCasting();
+		return;
+	}
+
 	g_chat->removeUserFromChannel(*player, channelId);
 }
 
@@ -3036,6 +3041,10 @@ void Game::playerSetFightModes(uint32_t playerId, fightMode_t fightMode, bool ch
 	player->setFightMode(fightMode);
 	player->setChaseMode(chaseMode);
 	player->setSecureMode(secureMode);
+
+	if (player->isLiveCasting()) {
+		player->sendFightModes(); // to update the spectator
+	}
 }
 
 void Game::playerRequestAddVip(uint32_t playerId, const std::string& name)
@@ -3233,7 +3242,7 @@ void Game::playerSay(uint32_t playerId, uint16_t channelId, SpeakClasses type,
 		return;
 	}
 
-	if (playerSaySpell(player, type, text)) {
+	if (playerSaySpell(player, type, channelId, text)) {
 		return;
 	}
 
@@ -3243,6 +3252,11 @@ void Game::playerSay(uint32_t playerId, uint16_t channelId, SpeakClasses type,
 
 	if (type != TALKTYPE_PRIVATE_PN) {
 		player->removeMessageBuffer();
+	}
+
+	if (channelId == CHANNEL_CAST) {
+		player->sendChannelMessage(player->getName(), text, TALKTYPE_CHANNEL_O, CHANNEL_CAST);
+		return;
 	}
 
 	switch (type) {
@@ -3282,11 +3296,11 @@ void Game::playerSay(uint32_t playerId, uint16_t channelId, SpeakClasses type,
 	}
 }
 
-bool Game::playerSaySpell(Player* player, SpeakClasses type, const std::string& text)
+bool Game::playerSaySpell(Player* player, SpeakClasses type, uint16_t channelId, const std::string& text)
 {
 	std::string words = text;
 
-	TalkActionResult_t result = g_talkActions->playerSaySpell(player, type, words);
+	TalkActionResult_t result = g_talkActions->playerSaySpell(player, type, channelId, words);
 	if (result == TALKACTION_BREAK) {
 		return true;
 	}
