+++ src/luascript.cpp
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -2290,6 +2290,16 @@ void LuaScriptInterface::registerFunctions()
 	registerMethod("Player", "hasChaseMode", LuaScriptInterface::luaPlayerHasChaseMode);
 	registerMethod("Player", "hasSecureMode", LuaScriptInterface::luaPlayerHasSecureMode);
 	registerMethod("Player", "getFightMode", LuaScriptInterface::luaPlayerGetFightMode);
+
+	registerMethod("Player", "isLiveCasting", LuaScriptInterface::luaPlayerIsLiveCasting);
+	registerMethod("Player", "startLiveCasting", LuaScriptInterface::luaPlayerStartLiveCasting);
+	registerMethod("Player", "pauseLiveCasting", LuaScriptInterface::luaPlayerPauseLiveCasting);
+	registerMethod("Player", "stopLiveCasting", LuaScriptInterface::luaPlayerStopLiveCasting);
+	registerMethod("Player", "kickCastSpectator", LuaScriptInterface::luaPlayerKickCastSpectator);
+	registerMethod("Player", "banCastSpectator", LuaScriptInterface::luaPlayerBanCastSpectator);
+	registerMethod("Player", "unBanCastSpectator", LuaScriptInterface::luaPlayerUnBanCastSpectator);
+	registerMethod("Player", "muteCastSpectator", LuaScriptInterface::luaPlayerMuteCastSpectator);
+	registerMethod("Player", "unMuteCastSpectator", LuaScriptInterface::luaPlayerUnMuteCastSpectator);
 
 	// Monster
 	registerClass("Monster", "Creature", LuaScriptInterface::luaMonsterCreate);
@@ -9216,6 +9226,8 @@ int LuaScriptInterface::luaPlayerSetGhostMode(lua_State* L)
 				it.second->notifyStatusChange(player, VIPSTATUS_OFFLINE);
 			}
 		}
+
+		player->stopLiveCasting();
 		IOLoginData::updateOnlineStatus(player->getGUID(), false);
 	} else {
 		for (const auto& it : g_game.getPlayers()) {
@@ -9347,6 +9359,124 @@ int LuaScriptInterface::luaPlayerGetFightMode(lua_State* L)
 	Player* player = getUserdata<Player>(L, 1);
 	if (player) {
 		lua_pushnumber(L, player->fightMode);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerIsLiveCasting(lua_State* L)
+{
+	// player:isLiveCasting()
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->isLiveCasting());
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerStartLiveCasting(lua_State* L)
+{
+	// player:startLiveCasting([password])
+	std::string password = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->startLiveCasting(password));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerPauseLiveCasting(lua_State* L)
+{
+	// player:pauseLiveCasting([reason])
+	std::string reason = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->pauseLiveCasting(reason);
+		lua_pushboolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerStopLiveCasting(lua_State* L)
+{
+	// player:stopLiveCasting()
+	std::string reason = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		player->stopLiveCasting();
+		lua_pushboolean(L, true);
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerKickCastSpectator(lua_State* L)
+{
+	// player:kickCastSpectator(name)
+	std::string name = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->kickCastSpectator(name));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerBanCastSpectator(lua_State* L)
+{
+	// player:banCastSpectator(name)
+	std::string name = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->banCastSpectator(name));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerUnBanCastSpectator(lua_State* L)
+{
+	// player:unBanCastSpectator(name)
+	std::string name = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->unBanCastSpectator(name));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerMuteCastSpectator(lua_State* L)
+{
+	// player:muteCastSpectator(name)
+	std::string name = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->muteCastSpectator(name));
+	} else {
+		lua_pushnil(L);
+	}
+	return 1;
+}
+
+int LuaScriptInterface::luaPlayerUnMuteCastSpectator(lua_State* L)
+{
+	// player:unMuteCastSpectator(name)
+	std::string name = getString(L, 2);
+	Player* player = getUserdata<Player>(L, 1);
+	if (player) {
+		lua_pushboolean(L, player->unMuteCastSpectator(name));
 	} else {
 		lua_pushnil(L);
 	}
