+++ src/player.h
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -253,6 +253,10 @@ class Player final : public Creature, public Cylinder
 
 		const GuildWarVector& getGuildWarVector() const {
 			return guildWarVector;
+		}
+
+		const std::map<uint8_t, OpenContainer>& getOpenContainers() const {
+			return openContainers;
 		}
 
 		Vocation* getVocation() const {
@@ -1097,6 +1101,38 @@ class Player final : public Creature, public Cylinder
 				client->writeToOutputBuffer(message);
 			}
 		}
+		bool isLiveCasting() {
+			return client && client->castinfo.enabled;
+		}
+		bool startLiveCasting(std::string password) {
+			return client && client->startLiveCasting(password);
+		}
+		void stopLiveCasting() {
+			if (client) {
+				client->stopLiveCasting();
+			}
+		}
+		void pauseLiveCasting(std::string reason) {
+			if (client) {
+				client->pauseLiveCasting(reason);
+			}
+		}
+		// cast methods;
+		bool kickCastSpectator(std::string name) {
+			return client && client->kickCastSpectator(name);
+		}
+		bool banCastSpectator(std::string name) {
+			return client && client->banCastSpectator(name);
+		}
+		bool unBanCastSpectator(std::string name) {
+			return client && client->unBanCastSpectator(name);
+		}
+		bool muteCastSpectator(std::string name) {
+			return client && client->muteCastSpectator(name);
+		}
+		bool unMuteCastSpectator(std::string name) {
+			return client && client->unMuteCastSpectator(name);
+		}
 
 		void receivePing() {
 			lastPong = OTSYS_TIME();
@@ -1333,6 +1369,8 @@ class Player final : public Creature, public Cylinder
 		friend class Actions;
 		friend class IOLoginData;
 		friend class ProtocolGame;
+		friend class ProtocolSpectator;
+		friend class ProtocolGameBase;
 };
 
 #endif
