+++ src/protocolgamebase.h
@@ -1,0 +1,246 @@
+﻿/**
+ * The Forgotten Server - a free and open-source MMORPG server emulator
+ * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+#ifndef FS_PROTOCOLGAMEBASE_H_FACA2A2D1A9348B78E8FD7E8003EBB87
+#define FS_PROTOCOLGAMEBASE_H_FACA2A2D1A9348B78E8FD7E8003EBB87
+
+#include "protocol.h"
+#include "chat.h"
+#include "creature.h"
+#include "tasks.h"
+
+class NetworkMessage;
+class Player;
+class Game;
+class House;
+class Container;
+class Tile;
+class Connection;
+class Quest;
+class ProtocolGame;
+using ProtocolGame_ptr = std::shared_ptr<ProtocolGame>;
+class ProtocolSpectator;
+using ProtocolSpectator_ptr = std::shared_ptr<ProtocolSpectator>;
+
+extern Game g_game;
+
+struct TextMessage
+{
+	MessageClasses type = MESSAGE_STATUS_DEFAULT;
+	std::string text;
+	Position position;
+	uint16_t channelId;
+	struct {
+		int32_t value = 0;
+		TextColor_t color;
+	} primary, secondary;
+
+	TextMessage() = default;
+	TextMessage(MessageClasses type, std::string text) : type(type), text(std::move(text)) {}
+};
+
+class ProtocolGameBase : public Protocol
+{
+	public:
+		// static protocol information
+		enum {server_sends_first = true};
+		enum {protocol_identifier = 0}; // Not required as we send first
+		enum {use_checksum = true};
+
+		explicit ProtocolGameBase(Connection_ptr connection) : Protocol(connection) {}
+
+		uint16_t getVersion() const {
+			return version;
+		}
+
+	private:
+		void checkCreatureAsKnown(uint32_t id, bool& known, uint32_t& removedKnown);
+
+		bool canSee(int32_t x, int32_t y, int32_t z) const;
+		bool canSee(const Creature*) const;
+		bool canSee(const Position& pos) const;
+
+		void disconnectClient(const std::string& message) const;
+		void onConnect() final;
+
+		virtual void writeToOutputBuffer(const NetworkMessage& msg, bool broadcast = true) = 0;
+
+		//Send functions
+		void sendChannelMessage(const std::string& author, const std::string& text, SpeakClasses type, uint16_t channel, bool broadcast = true);
+		void sendChannelEvent(uint16_t channelId, const std::string& playerName, ChannelEvent_t channelEvent, bool broadcast = true);
+		void sendClosePrivate(uint16_t channelId);
+		void sendCreatePrivateChannel(uint16_t channelId, const std::string& channelName);
+		void sendChannelsDialog();
+		void sendChannel(uint16_t channelId, const std::string& channelName, const UsersMap* channelUsers, const InvitedMap* invitedUsers);
+		void sendOpenPrivateChannel(const std::string& receiver);
+		void sendToChannel(const Creature* creature, SpeakClasses type, const std::string& text, uint16_t channelId);
+		void sendPrivateMessage(const Player* speaker, SpeakClasses type, const std::string& text);
+		void sendIcons(uint16_t icons);
+		void sendFYIBox(const std::string& message);
+
+		void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
+		void sendMagicEffect(const Position& pos, uint8_t type);
+		void sendCreatureHealth(const Creature* creature);
+		void sendSkills();
+		void sendPing();
+		void sendPingBack();
+		void sendCreatureTurn(const Creature* creature, uint32_t stackpos);
+		void sendCreatureSay(const Creature* creature, SpeakClasses type, const std::string& text, const Position* pos = nullptr);
+
+		void sendQuestLog();
+		void sendQuestLine(const Quest* quest);
+
+		void sendCancelWalk();
+		void sendChangeSpeed(const Creature* creature, uint32_t speed);
+		void sendCancelTarget();
+		void sendCreatureOutfit(const Creature* creature, const Outfit_t& outfit);
+		void sendStats();
+		void sendBasicData();
+		void sendTextMessage(const TextMessage& message);
+		void sendReLoginWindow(uint8_t unfairFightReduction);
+
+		void sendTutorial(uint8_t tutorialId);
+		void sendAddMarker(const Position& pos, uint8_t markType, const std::string& desc);
+
+		void sendCreatureWalkthrough(const Creature* creature, bool walkthrough);
+		void sendCreatureShield(const Creature* creature);
+		void sendCreatureSkull(const Creature* creature);
+		void sendCreatureType(uint32_t creatureId, uint8_t creatureType);
+		void sendCreatureHelpers(uint32_t creatureId, uint16_t helpers);
+
+		void sendShop(Npc* npc, const ShopInfoList& itemList);
+		void sendCloseShop();
+		void sendSaleItemList(const std::list<ShopInfo>& shop);
+		void sendMarketEnter(uint32_t depotId);
+		void sendMarketLeave();
+		void sendMarketBrowseItem(uint16_t itemId, const MarketOfferList& buyOffers, const MarketOfferList& sellOffers);
+		void sendMarketAcceptOffer(const MarketOfferEx& offer);
+		void sendMarketBrowseOwnOffers(const MarketOfferList& buyOffers, const MarketOfferList& sellOffers);
+		void sendMarketCancelOffer(const MarketOfferEx& offer);
+		void sendMarketBrowseOwnHistory(const HistoryMarketOfferList& buyOffers, const HistoryMarketOfferList& sellOffers);
+		void sendMarketDetail(uint16_t itemId);
+		void sendTradeItemRequest(const std::string& traderName, const Item* item, bool ack);
+		void sendCloseTrade();
+
+		void sendTextWindow(uint32_t windowTextId, Item* item, uint16_t maxlen, bool canWrite);
+		void sendTextWindow(uint32_t windowTextId, uint32_t itemId, const std::string& text);
+		void sendHouseWindow(uint32_t windowTextId, const std::string& text);
+		void sendOutfitWindow();
+
+		void sendUpdatedVIPStatus(uint32_t guid, VipStatus_t newStatus);
+		void sendVIP(uint32_t guid, const std::string& name, const std::string& description, uint32_t icon, bool notify, VipStatus_t status);
+		void sendVIPEntries();
+
+		void sendPendingStateEntered();
+		void sendEnterWorld();
+
+		void sendFightModes();
+
+		void sendCreatureLight(const Creature* creature);
+		void sendWorldLight(const LightInfo& lightInfo);
+
+		void sendCreatureSquare(const Creature* creature, SquareColor_t color);
+
+		void sendSpellCooldown(uint8_t spellId, uint32_t time);
+		void sendSpellGroupCooldown(SpellGroup_t groupId, uint32_t time);
+
+		//tiles
+		void sendMapDescription(const Position& pos);
+
+		void sendAddTileItem(const Position& pos, uint32_t stackpos, const Item* item);
+		void sendUpdateTileItem(const Position& pos, uint32_t stackpos, const Item* item);
+		void sendRemoveTileThing(const Position& pos, uint32_t stackpos);
+		void sendUpdateTile(const Tile* tile, const Position& pos);
+
+		void sendAddCreature(const Creature* creature, const Position& pos, int32_t stackpos, bool isLogin);
+		void sendMoveCreature(const Creature* creature, const Position& newPos, int32_t newStackPos,
+		                      const Position& oldPos, int32_t oldStackPos, bool teleport);
+
+		//containers
+		void sendAddContainerItem(uint8_t cid, uint16_t slot, const Item* item);
+		void sendUpdateContainerItem(uint8_t cid, uint16_t slot, const Item* item);
+		void sendRemoveContainerItem(uint8_t cid, uint16_t slot, const Item* lastItem);
+
+		void sendContainer(uint8_t cid, const Container* container, bool hasParent, uint16_t firstIndex);
+		void sendCloseContainer(uint8_t cid);
+
+		//inventory
+		void sendInventoryItem(slots_t slot, const Item* item);
+
+		//messages
+		void sendModalWindow(const ModalWindow& modalWindow);
+
+		//Help functions
+
+		// translate a tile to clientreadable format
+		void GetTileDescription(const Tile* tile, NetworkMessage& msg);
+
+		// translate a floor to clientreadable format
+		void GetFloorDescription(NetworkMessage& msg, int32_t x, int32_t y, int32_t z,
+		                         int32_t width, int32_t height, int32_t offset, int32_t& skip);
+
+		// translate a map area to clientreadable format
+		void GetMapDescription(int32_t x, int32_t y, int32_t z,
+		                       int32_t width, int32_t height, NetworkMessage& msg);
+
+		void AddCreature(NetworkMessage& msg, const Creature* creature, bool known, uint32_t remove);
+		void AddPlayerStats(NetworkMessage& msg);
+		void AddOutfit(NetworkMessage& msg, const Outfit_t& outfit);
+		void AddPlayerSkills(NetworkMessage& msg);
+		void AddWorldLight(NetworkMessage& msg, const LightInfo& lightInfo);
+		void AddCreatureLight(NetworkMessage& msg, const Creature* creature);
+
+		//tiles
+		static void RemoveTileThing(NetworkMessage& msg, const Position& pos, uint32_t stackpos);
+
+		void MoveUpCreature(NetworkMessage& msg, const Creature* creature, const Position& newPos, const Position& oldPos);
+		void MoveDownCreature(NetworkMessage& msg, const Creature* creature, const Position& newPos, const Position& oldPos);
+
+		//shop
+		void AddShopItem(NetworkMessage& msg, const ShopInfo& item);
+
+		friend class Player;
+		friend class ProtocolGame;
+		friend class ProtocolSpectator;
+
+		// Helpers so we don't need to bind every time
+		template <typename Callable, typename... Args>
+		void addGameTask(Callable function, Args&&... args) {
+			g_dispatcher.addTask(createTask(std::bind(function, &g_game, std::forward<Args>(args)...)));
+		}
+
+		template <typename Callable, typename... Args>
+		void addGameTaskTimed(uint32_t delay, Callable function, Args&&... args) {
+			g_dispatcher.addTask(createTask(delay, std::bind(function, &g_game, std::forward<Args>(args)...)));
+		}
+
+		std::unordered_set<uint32_t> knownCreatureSet;
+		Player* player = nullptr;
+
+		uint32_t eventConnect = 0;
+		uint32_t challengeTimestamp = 0;
+		uint16_t version = CLIENT_VERSION_MIN;
+
+		uint8_t challengeRandom = 0;
+
+		bool debugAssertSent = false;
+		bool acceptPackets = false;
+};
+
+#endif
