+++ src/talkaction.cpp
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -66,10 +66,14 @@ bool TalkActions::registerEvent(Event* event, const pugi::xml_node&)
 	return true;
 }
 
-TalkActionResult_t TalkActions::playerSaySpell(Player* player, SpeakClasses type, const std::string& words) const
+TalkActionResult_t TalkActions::playerSaySpell(Player* player, SpeakClasses type, uint16_t channelId, const std::string& words) const
 {
 	size_t wordsLength = words.length();
 	for (const TalkAction& talkAction : talkActions) {
+		if (talkAction.getChannel() != 0 && channelId != talkAction.getChannel()) {
+			continue;
+		}
+		
 		const std::string& talkactionWords = talkAction.getWords();
 		size_t talkactionLength = talkactionWords.length();
 		if (wordsLength < talkactionLength || strncasecmp(words.c_str(), talkactionWords.c_str(), talkactionLength) != 0) {
@@ -116,6 +120,17 @@ bool TalkAction::configureEvent(const pugi::xml_node& node)
 	pugi::xml_attribute separatorAttribute = node.attribute("separator");
 	if (separatorAttribute) {
 		separator = pugi::cast<char>(separatorAttribute.value());
+	}
+
+	pugi::xml_attribute channelAttibute = node.attribute("channel");
+	if (channelAttibute) {
+		std::string channelName = asLowerCaseString(channelAttibute.as_string());
+		if (channelName == "cast") {
+			channel = CHANNEL_CAST;
+		} else {
+			std::cout << "[Error - TalkAction::configureEvent] Unknown channel name: " << channelName << std::endl;
+			return false;
+		}
 	}
 
 	words = wordsAttribute.as_string();
