+++ src/talkaction.h
@@ -1,4 +1,4 @@
-/**
+﻿/**
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2017  Mark Samman <mark.samman@gmail.com>
  *
@@ -43,6 +43,9 @@ class TalkAction : public Event
 		char getSeparator() const {
 			return separator;
 		}
+		uint16_t getChannel() const {
+			return channel;
+		}
 
 		//scripting
 		bool executeSay(Player* player, const std::string& param, SpeakClasses type) const;
@@ -53,6 +56,7 @@ class TalkAction : public Event
 
 		std::string words;
 		char separator = '"';
+		uint16_t channel = 0;
 };
 
 class TalkActions : public BaseEvents
@@ -65,7 +69,7 @@ class TalkActions : public BaseEvents
 		TalkActions(const TalkActions&) = delete;
 		TalkActions& operator=(const TalkActions&) = delete;
 
-		TalkActionResult_t playerSaySpell(Player* player, SpeakClasses type, const std::string& words) const;
+		TalkActionResult_t playerSaySpell(Player* player, SpeakClasses type, uint16_t channelId, const std::string& words) const;
 
 	protected:
 		LuaScriptInterface& getScriptInterface() final;
